fix: web / desktop redirect
