+++ src/App.svelte
@@ -20,20 +20,6 @@
 	export let url = '/';
 
 
-
-	onMount(() => {
-		// Check the query parameters in the URL
-		const urlParams = new URLSearchParams(window.location.search);
-		const source = urlParams.get('source'); // Get 'source' from query param
-		if(!source) return;
-		if (source === 'web') {
-			localStorage.setItem('isUserFromDesktop', 'false'); 
-			
-		} else  {
-			localStorage.setItem('isUserFromDesktop', 'true'); 
-		} 
-	});
-
 </script>
 
 <Router {url}>
